requirements.txt
deepface==0.0.79
keras==2.13.1
matplotlib==3.7.1
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
py_feat==0.6.0
